run_make_support: move artifact name helpers into `artifact_names` module
